Add support for saving and loading weights
A model with group/id will save its weights to `weights/group/id/id-val.h5'
after each epoch and will save the weight corresponding to the best accuracy to
`weights/group/id/id-acc.h5'. If you want to have the model pick up where it
left off, then just run the experiment again and the model will see that
`weights/group/id/id-val.h5` exists, load the weights, and start training.

One hack I had to make was making it so whatever string you pass as `-data-dir`
gets prepended with `pickle_jar`. This is because the exp generation script uses
sed to generate the exp files and sed doesn't like slashes passed to it. This is
obviously problematic if we want to group train and test images into nested
directories within pickle_jar, so we'll have to adjust accordingly in the
future.